Problem: users can claim & unclaim issues from home route #61
Solution: Move actions to details page
