+++ imports/ui/components/documents/index/documents-index-item/documents-index-item.html
@@ -6,10 +6,8 @@
     <td>{{showTimeAgo document.updatedAt}}</td>
     <td>{{showPrettyTimestamp document.createdAt}}</td>
     <td class="text-right">
-      {{{claimButton}}}
 	  {{#if isProblemOwner document.createdBy}}
 		<a class="btn btn-sm btn-secondary" href="{{pathFor 'documentEdit' documentId=document._id}}" role="button">Edit</a>
-		<a class="btn btn-sm btn-danger js-delete-document" href="#" role="button">Delete</a>
 	  {{/if}}
     </td>
   </tr>
