+++ imports/ui/components/documents/show/document-show.html
@@ -4,18 +4,17 @@
     <div class="card">
       <div class="card-body">
         {{#if Template.subscriptionsReady}}
+          <div class="float-right">
+            {{{claimButton problem}}}
+            {{#if isProblemOwner problem.createdBy}}
+        		<a class="btn btn-sm btn-secondary" href="{{pathFor 'documentEdit' documentId=problem._id}}" role="button">Edit</a>
+        		<a class="btn btn-sm btn-danger js-delete-document" href="#" role="button">Delete</a>
+        	  {{/if}}
+          </div>
           {{#if problem.claimed}}
           <span class="badge badge-success">CLAIMED</span>
           {{/if}}
-          
-
-          {{#if isProblemOwner problem.createdBy}}
-            <a class="btn btn-sm btn-secondary pull-right" href="/{{problem._id}}/edit">Edit</a>
-          {{/if}}
-          
-          <h2>
-            {{problem.summary}}
-          </h2>
+          <h2>{{problem.summary}}</h2>
           <p>{{problem.description}}</p>
           <hr>
           <p class="text-muted">
@@ -24,7 +23,7 @@
             {{/if}}
 
           {{#if problem.claimed}}
-          <strong>Claimed by {{problem.claimedFullname}} {{showTimeAgoTimestamp problem.claimedDateTime}}</strong> <br /> 
+          <strong>Claimed by {{problem.claimedFullname}} {{showTimeAgoTimestamp problem.claimedDateTime}}</strong> <br />
           {{/if}}
           Created: {{showPrettyTimestamp problem.createdAt}}
           </p>
